refactor: announcement-to-notice
