+++ app/src/main/java/com/saegil/android/navigation/NavGraph.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.saegil.announcement.announcement.NoticeScreen
+import com.saegil.notice.notice.NoticeScreen
 import com.saegil.learning.learning.LearningScreen
 import com.saegil.map.map.MapScreen
 import com.saegil.mypage.mypage.MypageScreen
